% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_bmi.R
\name{cal_bmi}
\alias{cal_bmi}
\title{cal_bmi}
\usage{
cal_bmi(height, weight)
}
\arguments{
\item{height}{The height of the participants}

\item{weight}{The weight of the participants}
}
\value{
BMI
}
\description{
Calculate the BMI of participants
}
\details{
This function is used to calculate the BMI for participants.
}
\author{
Yi-Han Tang
}
