% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_check.R
\name{check_check}
\alias{check_check}
\title{Title check_check}
\usage{
check_check(age, height, weight)
}
\arguments{
\item{age}{The age of the participants}

\item{height}{The height of the participants}

\item{weight}{The weight of the participants}
}
\value{

}
\description{
This function is used to check the qualify of participants.
}
\details{
This function is used to check the qualify of participants in order to find the target participants.
The target participants are the people whose age above 20 years-old and BMI is equal or higher than 25.
}
\examples{
check_check(25, 170,100)
"Eligile"

check_check(19,170,100)
"Ineligible"

}
\author{
Yi-Han Tang
}
