% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgender.R
\name{plotgender}
\alias{plotgender}
\title{plotgender}
\usage{
plotgender(age, height, weight, ...)
}
\arguments{
\item{age}{The age of the participants}

\item{height}{The height of the participants}

\item{weight}{The weight of the participants}

\item{data}{data should be a data frame which should contain BMI, age, and drug_frequency}
}
\value{
plot
}
\description{
This function is used to plot the participants by different variable
}
\details{
This function is used to see the BMI distribution of participants of different ages.
Moreover, the participants are group by how many times they are taking the pills in a week.
}
\author{
Yi-Han Tang
}
