% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgender_separate.R
\name{plotgender_separate}
\alias{plotgender_separate}
\title{plotgender_separate}
\usage{
plotgender_separate(data, age, height, weight, gender, ...)
}
\arguments{
\item{data}{data should be a data frame which should contain BMI, age, and drug_frequency}

\item{age}{The age of the participants}

\item{height}{The height of the participants}

\item{weight}{The weight of the participants}

\item{gender}{}
}
\value{
plot
}
\description{
This function is used to plot the participants by different varies
}
\details{
This function is used to see the BMI distribution of participants of different ages.
Moreover, the participants are group by how many times they are taking the pills in a week.
}
\examples{

}
\author{
Yi-Han Tang
}
